fix: move some types to debug protocol
