+++ packages/origine2/src/types/debugProtocol.ts
@@ -29,3 +29,23 @@ export interface IDebugMessage {
     stageSyncMsg: IStageState;
   };
 }
+
+export interface IComponentsVisibility {
+  showStarter: boolean; // 是否显示初始界面（用于使得bgm可以播放)
+  showTitle: boolean; // 是否显示标题界面
+  showMenuPanel: boolean; // 是否显示Menu界面
+  showTextBox: boolean;
+  showControls: boolean;
+  controlsVisibility: boolean;
+  showBacklog: boolean;
+  showExtra: boolean;
+  showGlobalDialog: boolean;
+  showPanicOverlay: boolean;
+  isEnterGame: boolean;
+  isShowLogo: boolean;
+}
+
+export interface IComponentVisibilityCommand {
+  component: keyof IComponentsVisibility;
+  visibility: boolean;
+}
